Refactoring to services
